[3] - Modelo base para rotas, decorators e muito mais (#13)
* Melhorias na infra basica

* Refatoração de nomes e linter

* Teste linter

* FIX - linter

* WIP - configuração do banco de dados

* WIP - primeiro modelo e testes de migracao com alembic

* WIP - testando listagem de registros

* WIP - modelo base para rotas e decorators

* WIP - adições a infra do projeto

* WIP - adicionando swagger

* FIX - correções nos models

* FIX - correções no modelo de rotas e listener para sqlalchemy

+++ aplicativo/models/catalogo/Catalogo.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, BigInterger, ForeignKey
+from sqlalchemy import Column, BigInteger, String, ForeignKey
 from pydantic import BaseModel, constr
 from aplicativo.database import Base
 
 
 class Catalogo(Base):
     __tablename__ = "catalogo"
-    id = Column(Integer, primary_key=True, nullable=False)
+    id = Column(BigInteger, primary_key=True, nullable=False)
     nome = Column(String(255))
     perfil_id = Column(ForeignKey("perfil.id"), nullable=False)
     descricao = Column(String(255), nullable=False)
