+++ aplicativo/models/categoria/Categoria.py
@@ -1,11 +1,11 @@
-from sqlalchemy import Column, Integer, String, ForeignKey
+from sqlalchemy import Column, BigInteger, String, ForeignKey
 from pydantic import BaseModel, constr
 from aplicativo.database import Base
 
 
 class Categoria(Base):
     __tablename__ = "categoria"
-    id = Column(Integer, primary_key=True, nullable=False)
+    id = Column(BigInteger, primary_key=True, nullable=False)
     nome = Column(String(255))
     descricao = Column(String(255), nullable=False)
     perfil_id = Column(ForeignKey("perfil.id"), nullable=False)
